Fix breaking Avalonia changes
