+++ Apollo/DeviceViewers/GroupViewer.cs
@@ -215,7 +215,7 @@ namespace Apollo.DeviceViewers {
                 return;
             }
 
-            bool copy = e.Modifiers.HasFlag(Program.ControlKey);
+            bool copy = e.Modifiers.HasFlag(App.ControlKey);
             bool result;
 
             if (e.Data.Contains("chain")) {
