+++ Apollo/Helpers/SelectionManager.cs
@@ -126,7 +126,7 @@ namespace Apollo.Helpers {
         public bool HandleKey(KeyEventArgs e) {
             if (Start == null) return true;
 
-            if (e.Modifiers == Program.ControlKey) {
+            if (e.Modifiers == App.ControlKey) {
                 if (e.Key == Key.X) Action("Cut");
                 else if (e.Key == Key.C) Action("Copy");
                 else if (e.Key == Key.D) Action("Duplicate");
