+++ Apollo/Viewers/ChainViewer.cs
@@ -185,7 +185,7 @@ namespace Apollo.Viewers {
             Chain source_parent = moving[0].Parent;
             int before = moving[0].IParentIndex.Value - 1;
 
-            bool copy = e.Modifiers.HasFlag(Program.ControlKey);
+            bool copy = e.Modifiers.HasFlag(App.ControlKey);
 
             bool result = Device.Move(moving, _chain, after, copy);
