+++ Apollo/Viewers/TrackInfo.cs
@@ -191,7 +191,7 @@ namespace Apollo.Viewers {
 
             List<Track> moving = ((List<ISelect>)e.Data.Get("track")).Select(i => (Track)i).ToList();
             int before = moving[0].IParentIndex.Value - 1;
-            bool copy = e.Modifiers.HasFlag(Program.ControlKey);
+            bool copy = e.Modifiers.HasFlag(App.ControlKey);
 
             bool result = Track.Move(moving, Program.Project, after, copy);
