+++ Apollo/Windows/SplashWindow.cs
@@ -115,18 +115,16 @@ namespace Apollo.Windows {
 
             if (Launchpad.CFWIncompatible == CFWIncompatibleState.Show) Launchpad.CFWError(this);
 
-            if (Program.Args?.Length > 0)
-                ReadFile(Program.Args[0]);
+            if (App.Args?.Length > 0)
+                ReadFile(App.Args[0]);
             
-            Program.Args = null;
+            App.Args = null;
 
             UpdateBlogpost();
             UpdateRelease();
 
             if (IsVisible && !openDialog && await Github.ShouldUpdate()) {
-                Window[] windows = Application.Current.Windows.ToArray();
-                
-                foreach (Window window in windows)
+                foreach (Window window in App.Windows)
                     if (window.GetType() != typeof(MessageWindow))
                         window.Close();
                 
@@ -142,7 +140,7 @@ namespace Apollo.Windows {
 
             this.Content = null;
 
-            Program.WindowClosed(this);
+            App.WindowClosed(this);
         }
 
         void TabChanged(int tab) {
@@ -151,7 +149,7 @@ namespace Apollo.Windows {
                     RecentProjectInfo viewer = new RecentProjectInfo(Preferences.Recents[i]);
                     viewer.Opened += ReadFile;
                     viewer.Removed += Remove;
-                    viewer.Showed += Program.URL;
+                    viewer.Showed += App.URL;
 
                     Recents.Children.Add(viewer);
                 }
@@ -248,10 +246,10 @@ namespace Apollo.Windows {
         }
 
         async void Blogpost(object sender, RoutedEventArgs e)
-            => Program.URL($"https://apollo.mat1jaczyyy.com/post/{Path.GetFileNameWithoutExtension((await Github.LatestBlogpost()).Name)}");
+            => App.URL($"https://apollo.mat1jaczyyy.com/post/{Path.GetFileNameWithoutExtension((await Github.LatestBlogpost()).Name)}");
 
         async void Release(object sender, RoutedEventArgs e)
-            => Program.URL((await Github.LatestRelease()).HtmlUrl);
+            => App.URL((await Github.LatestRelease()).HtmlUrl);
 
         void Restore(object sender, RoutedEventArgs e) {
             CrashPanel.Opacity = 0;
@@ -277,9 +275,9 @@ namespace Apollo.Windows {
         }
 
         void Window_KeyDown(object sender, KeyEventArgs e) {
-            if (Program.WindowKey(this, e)) return;
+            if (App.WindowKey(this, e)) return;
 
-            if (e.Modifiers == Program.ControlKey) {
+            if (e.Modifiers == App.ControlKey) {
                 if (e.Key == Key.N) New(sender, e);
                 else if (e.Key == Key.O) Open(sender, e);
             }
